+++ mirtop/gff/body.py
@@ -187,6 +187,8 @@ def variant_with_nt(line, precursors, matures):
     mm = align_from_variants(read,
                              mature_sequence,
                              attr["Variant"])
+    if mm == "Invalid":
+        return mm
     if len(mm) > 0:
         mm = "".join(["".join(map(str, m)) for m in mm])
     else: