+++ test/test_automated_analysis.py
@@ -156,6 +156,25 @@ class AutomatedAnalysisTest(unittest.TestCase):
             print " ".join(clcode)
             subprocess.check_call(clcode)
 
+    @attr(complete=True)
+    @attr(cmd_srnabench=True)
+    def test_srnaseq_annotation_srnabench(self):
+        """Run srnabench analysis
+        """
+        with make_workdir():
+            clcode = ["mirtop",
+                      "gff",
+                      "--format", "srnabench",
+                      "--sps", "hsa",
+                      "--hairpin", "../../data/examples/annotate/hairpin.fa",
+                      "--gtf", "../../data/examples/annotate/hsa.gff3",
+                      "-o", "test_out_mirs",
+                      "../../data/examples/srnabench",
+                      "-d", "-vd"]
+            print ""
+            print " ".join(clcode)
+            subprocess.check_call(clcode)
+
 
     @attr(complete=True)
     @attr(cmd_stats=True)